add slash command for messaging

+++ commands/refresh.js
@@ -6,9 +6,15 @@ module.exports = {
 		.setName('refresh')
 		.setDescription('Removes all stored nicknames and colors!'),
 	async execute(interaction) {
-		fs.unlinkSync('../files/Nicknames.txt')
-		fs.appendFileSync('./files/Nicknames.txt',`\n`)
-		
-		await interaction.reply({ content: 'I have cleared all nicknames!', components: [] });
+
+		// Check executor. If not Dylan, do nothing.
+		if (interaction.user.id !== 200316185445793792) {
+			await interaction.reply({ content: 'You\'re not <@200316185445793792>! You can\'t do that.', components: [] });
+		} else {
+			fs.unlinkSync('../files/Nicknames.txt');
+			fs.appendFileSync('./files/Nicknames.txt',`\n`);
+			
+			await interaction.reply({ content: 'I have cleared all nicknames!', components: [] });
+		};
 	},
 };
